+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `YAMLSimple`: added flow sequence
+
 ## [0.1] - 2020-10-05
 ### Added
 - all classes: initial versions